requirements.txt
Flask==3.0.3
flask_wtf==1.2.1
injector==0.21.0
openai==1.33.0
python-dotenv==1.0.1
WTForms==3.1.2
zhipuai==2.1.0.20240521
